test(doctor): isolate wizard metadata ownership
